+++ lib/CalculatorPage/Style.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+const inputTextMaxLength = 5;
+
+const TextStyle titleStyle = TextStyle(
+  fontSize: 22,
+  fontWeight: FontWeight.bold,
+);
+
+const TextStyle inputStyle = TextStyle(
+  fontSize: 18,
+);
+
+const TextStyle buttonStyle = TextStyle(
+  fontSize: 18,
+);
