+++ lib/TrackerPage.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:mybarbell/provider/currentSettings.dart';
 import 'package:provider/provider.dart';
+import 'provider/currentSettings.dart';
 
 class TrackerPage extends StatefulWidget {
   const TrackerPage({super.key});
