+++ lib/l10n/messages_en.arb
@@ -48,7 +48,7 @@
   "advanced": "Advanced",
   "elite": "Elite",
 
-
+  "one_rep_copied": "1RM Copied",
   "weight_not_specified": "Weight not specified",
   "reps_not_specified": "Reps not specified",
   "body_weight_not_specified": "Body weight not specified",
